Fixed max_token_channels to be u16.

+++ src/networker/messenger/handle_neighbor.rs
@@ -22,7 +22,7 @@ pub struct NeighborInconsistencyError {
 
 #[allow(unused)]
 pub struct NeighborSetMaxTokenChannels {
-    max_token_channels: u64,
+    max_token_channels: u16,
 }
 
 #[allow(unused)]
